Move types to a separated file
Move helpers to a directory
