+++ src/App.tsx
@@ -2,16 +2,8 @@ import { useQuery } from "@tanstack/react-query"
 import { api } from "./api/apiClient"
 import { StatusScreen } from "./components/StatusScreen"
 import { POSFlow } from "./features/pos/"
-import type { Schemas } from "./api/apiClient"
+import type { Stock } from "./types"
 
-export type Product = Schemas["ProductResponse"]
-export type Products = Product[]
-export type Salesman = Schemas["SalesmanResponse"]
-export type Salesmen = Salesman[]
-export type Stock = Record<string, number>
-export type PaymentType = Schemas["PaymentType"]
-export type SaleRequest = Schemas["SaleRequest"]
-export type SalesRequests = SaleRequest[]
 
 export default function App() {
     const {
