+++ src/features/pos/hooks/useCart.ts
@@ -1,5 +1,5 @@
 import { useState } from "react"
-import type { Products, Stock } from "@/App"
+import type { Products, Stock } from "@/types"
 
 export function useCart(products: Products, stock: Stock) {
     // Core state. Single source of truth. Simplest representation of the cart
