+++ src/types/index.ts
@@ -1,0 +1,10 @@
+import type { Schemas } from "../api/apiClient"
+
+export type Product = Schemas["ProductResponse"]
+export type Products = Product[]
+export type Salesman = Schemas["SalesmanResponse"]
+export type Salesmen = Salesman[]
+export type Stock = Record<string, number>
+export type PaymentType = Schemas["PaymentType"]
+export type SaleRequest = Schemas["SaleRequest"]
+export type SalesRequests = SaleRequest[]
